Start building IR (#20)
Intermediate Representation is needed to resolve a lot of odd
constraints that don't really belong in the parse phase. Examples:

* IR should figure out the correct names for cdk vs not cdk
* IR should figure out ordering issues, if they arise

Post-IR phases should take IR and expect essentially "just instructions"
to be plopped out from the IR in the language of choice. You should be
able to analyze IR and always know what will be emitted, which is also
helpful for tests.

Conditions is the first move to IR, though a lot of it is just duplicate
traversal of the parse tree, which is unfortunate.