Move all but resources to ir output (#24)
* Move all but resources to ir output

Semantics as an idea has been trasnformed to ir. All of the weird
off-by-one functionality that was there is now all in ir as a single
unit.

IR is now the "I see a parse tree and can make a lot of decisions
from it". It does the semantic-level business logic of combining
multiple areas of the parse tree to make a sensible, eventual output.

Most of parser was also carved out for formatting and output. That was
put into synthesizer instead. This feels more like a compiler now
with specific, and separate responsibilities and phases.