Add support for metadata and update policy (#37)
* Add metadata output

For things like autoscaling groups, you can add cfn init data.
This data is important to actually initialize the correct state
for those types of policies.

This is the first half for this. The second is "UpdatePolicy",
which will come at a later point.

* Add full UpdatePolicy support

UpdatePolicy is used for asgs / lambda functions for actually updating
in a safe way. Noctilucent now supports those translations along with
metadata, which means full ASG support now works.

* Add tests