Fix no substitution bug (#70)
There was a recent bug filed around substitution not working for
a "no substitution" corner case. The parse would not handle such
a case because it could not consume the rest of the output UNLESS
it had a $ in it.

As such, rewrote the parser a bit to consume the rest of the tokens
and return eof on the next go round instead. Also move Sub to IR.
Though it is technically a parser, the tokenization of the output
doesn't occur until IR output, so it belongs there more. It may
make sense to revisit if we want that level of tokenization in an
earlier phase in the future, but for now this seems fine.

Fixes #66