Update lookup table to handle arbitrary types
It turns out that cloudformation actually allows arbitrary types
in Fn::Mappings. Within that, it's also not uniform types as well.

This update handles the arbitrary types for non-array-based inputs
and will "opt out" for non-array-based complex structures with an
`any` mode, leaving it up to the user to resolve that problem in
typescript. Also, in order to reuse f64 types, moved WrapperF64
to a primitives folder that spans the gap of parser and ir.