Upgrade type structure for Fn::FindInMappings
When passing down types into an Fn::Mapping, I originally expected
to continue passing down the types into deeply nested structures.

This actually doesn't make sense. As an example, let's take a look
at ReadCapacityUnits in DDB:

```
"ReadCapacityUnits" : 3000
```

Simple. It requires an integer.

When reading this fun mutated version:

```
"ReadCapacityUnits": {"Fn::FindInMap": ["DDBValue", "Table"] }
```

or something or other. ReadCapacityUnits is indeed still requiring an
integer, but FindInMap is not, it is requiring two input strings in
an array. The original code would pass the complexity type from the
parent (in this case RSU), but that is incorrect and so would not work
for resolving anything other than strings in maps. In this newest iteration,
Mappings will now only resolve strings. This may cause chaos down the road,
but according to the spec, all mapping keys are strings, so I'll allow it
for now.