Fix deduping problems in specification file (#80)
The specification for CloudFormation sucks. It doesn't follow it's
own standards and creates invalid JSON. This change fixes some of
those rough edges by deduping in a hashmap and "hoping for the best".

Also change required type to be optional since some teams can't publish
correctly on that as well.