chore: bake in specification using a PHF lookup table
Instead of inlining the JSON spec in the binary, and applying a two-pass
transformation to turns that into the `Specification` object, we now use
build-time code-generation to create static PHF maps with the literal
values for the property & resource type specifications.

This removes the need for a run-time parsing pass, and allows the lookup
table to be statically built at compile time, with a very efficient
lookup method.

Additionally hid the properties of the `Specification` struct and
exposing the underlying data only through accessors.