feat: code-generate PHF lookup maps for the specification (#107)
Overhauls the modeling of the specification types to be more `enum`-heavy (this provides better semantics than a set of options that are contextually exclusive), and switch from including the complete JSON document of the specification in the binary to code-generating efficient lookup maps at build time. This has the effect of reducing the release binary size from ~11MiB to ~3.1MiB, which is quite an improvement (70% shrink).