feat: add support for `Fn::Split`
Added unit tests for parsing of all intrinsic functions as
`ResourceValue` objects and added missing support for the `Fn::Split`
instrinsic function.

Note that the `Fn::Transform` function is still unsupported, and neither
are the `Fn::Length` and `Fn::ToJsonString` functions which require the
`AWS::LanguageExtensions` transform to be used.