chore: Fix PR Action (#142)
Pull Requests for schema changes are no longer working. They started to break because in v1.5+, dist files were no longer included in the action -- so things are just busted. 

The recommendation both on the marketplace and the issue is to just use the recommended version (either v1.x or v1). We don't need advanced functionality, so tagging us to v1.