chore: improved code generation API
The previous `CodeSink` API immediately flushed out into the output
writer, hence preventing additional insertions into the previous code.
However, in order to avoid introducing unnecessary imports without
having to traverse the entire tree to decide whether a particular import
is necessary or not, lazy addition of new imports is desirable.

This introduces a new `CodeBuffer` struct which aggregates code in
memory while keeping track of indentation requirements, so that code can
be inserted back into previous sections of the file prior to it being
written.

This also has the effect of drastically reducing the amount of branches,
as there are far fewer failable calls involved when generating code now.