fix(go): use generics trick to implement Fn::If
Go does not have a ternary operator and if-else statements are not
expressions, so one needs to compose an IIFE in order to emulate the
semantics of the elvis operator (`?:`).

This can be pretty wordy, and requires being able to determine the
appropriate value type locally, which is cumbersome. Fortunately, go
generics can be used to let the Go compiler infer the correct type to
use, and this in turns reduces the verbosity of the construct.

This introduces an `ifCondition` generic function to implement the
ternary expression in a generic manner, which solves that problem.