chore: add release workflows for crates and npm (#245)
* chore: add release workflows for crates and npm

* add branch sync to release workflow

This change adds release-sync, which once per day will push any changes to the release branch.

On a push to that branch, the release workflow will increment the version, publish the new version to crates.io, tag the release commit, and then push the commit back to the release branch.

It then creates an auto-approve PR to push the release back to main. Lastly, the wasm-pack is built and released to npm.