chore: add security audits (#272)
This adds a check to each individual PR as well as a daily check. The checks on individual PRs will ensure we don't merge anything with changes that include bad dependencies. It will fail the workflow so it cannot be merged. The cron job will cut an issue on a failure, in the case of emergent security advisories for dependencies we're already using.