fix: GetAtt functions may include "." and cdk functions do not (#315)
* fix: GetAtt functions may include '.' and cdk functions do not

Line 245 in the template contains a Fn::GetAtt with 'Compliance.Type' and it is not handled correctly in any language.

The tests are being left out of this PR because not all issues have been resolved yet. The template is included for context.

* fix(java, csharp): new line not properly escaped in string join

This also fixes casing in one place for Java. These have also been checked against
https://github.com/cdklabs/cdk-from-cfn/pull/315/files#diff-16af47ca157647af1f02c784ec1666ee56accbcbbe84caccb8f1aa6533957f45
but the tests are not yet passing so I'm continuing the incremental bug fixes in the meantime.

* make the change in ir instead

* missed change