fix(csharp): import statements are incorrectly cased
This required quite a bit of refactoring since the default case we were using was for typescript.

We shouldn't be setting those defauts in IR so I removed it and handled the language specific logic in each language synthesizer.

This change legitimately fixes the issue for all use cases (or at least every one we have in the cdk library.