fix(csharp): import statements are incorrectly cased (#320)
* fix(csharp): import statments are incorrectly cased

This change makes all imports pascal cased if they are four letters or longer and in all caps if no.

* fix: GetAtt functions may include '.' and cdk functions do not

Line 245 in the template contains a Fn::GetAtt with 'Compliance.Type' and it is not handled correctly in any language.

The tests are being left out of this PR because not all issues have been resolved yet. The template is included for context.

* fine clippy have it your way

* fix(csharp): import statements are incorrectly cased

This required quite a bit of refactoring since the default case we were using was for typescript.

We shouldn't be setting those defauts in IR so I removed it and handled the language specific logic in each language synthesizer.

This change legitimately fixes the issue for all use cases (or at least every one we have in the cdk library.