fix: allow translation of doubles from strings (#751)
* fix: allow translation of doubles from strings

There were certain templates that errored out and said "not a valid
integer". When looking at the code, it turns out that if a string
thinks there is a number -- a number in JSON can be either integer
or doable. Originally, the code parsed only ints, so any float/double
fails outright. This update now will try both. I'm sure in the future
we will need to try to parse one, fail, parse the other -- but for
now, just look for a ".".

* Fix unit test

---------

Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>